requirements.txt
Flask
flask-cors
flask-pymongo
Werkzeug
PyJWT